Added the build script
Since the orgs/github/members endpoint does not give us name information, we have to query the name for each member. That means a lot of requests. This script will do it for us, but it needs a token (since there are many requests). So the usage is:

node build.js <token>

It will override the js/Hubbers.js file with the new information. On the client side, when the requests fail (due to the rate limit), it will fallback to this file information which contains all the members of the GitHub organization (fetched when the script was run).